Added functions to extract epub archive,fixed header
In this commit i added functionality to extract an epub archive to a temporary location.Again i tried to keep the design modular. This will be changed as changes come along.
I also fixed the header to the standard atril backend header. Performed some code cleaning, naming functions as in the atril style of coding.