Created a Makefile for minizip, modified epub Makefile
I created a Makefile for the minizip backend, since i am only using part of that library and modified the epub backend to work with the library.
Also modified the configure script to suit our purpose.
The epub backend can now be compiled using the make command in the epub  directory, although it cannot still be compiled as a part of atril.