Epub backend builds as a part of Atril compilation
Fixed a tabbing error in backend/Makefile.am due to which the epub backend was not a target when building recursilvely, i am not commiting the changes to
epub-document.c as currently it will break the build, hope to fix it and get webkit and mathjax done by the weekend.

Also fixed an error in configure.ac wherin the appearance of AM_LIBTOOL_PROG before AC_CXX_PROG caused a "link:unsupported hardcode properties error" on Linux Mint.