Program no longer segfaults, but does not give thumbnails
The loading icons for all thumbnails are displayed, but the program does not segfault. Also realized that the absurd resizing of the window was due to the ephy-zoom action on the webview window. Will fix that in the next commit. Will also think of some workaround for screenshots.

For now committed this code so that this can act as a clean slate if I screw up somewhere.