Reading epub documents possible
Alebit with a bit of bugs as far as navigation etc. are possible, we can read an epub document in Atril.
The document will be displayed with thumbnails, and all keyboard shortcuts work.
Menu options, the ones that I have currently added do work.

I Will now move on to features such as inverted colors, document index, etc, most of which should be simple to add.

Also made changes so that the presentation view does not fire up while displaying an epub document.

Also added the feature to copy text to clipboard using the Atril menu.