Window title thumbnail job fixed(ePub)
In the design change in the previous commit, I forgot to make the thumbnail job used to set the window title a main loop job, I fix that in this commit.

Epub thumbnail jobs running as threads, WILL cause a program error occassionally, because we need to use the offscreen webview.