libview: add a method to get the job currently running in the worker thread
When a job is cancelled while it's running, the cancelled signal might be
emitted before the job finishes, and since the finished signal is not
emitted for cancelled jobs, it's not possible to know when the job has
finished. With this method we can see whether the job is still running
and wait until it finishes.

Taken from evince commit: 23e76eac47c60ab885edcdb6a337ee7587afa7e8
From: Carlos Garcia Campos <carlosgc@gnome.org>