Add accelerators: fit-page, full-width, dual-page
Summary:
This commit adds support for accelerators `f` (fit page in window),
`w` (fill window width with page), and `d` (toggle dual-page display).
These accelerators are present in evince and are extremely commonly used
(at least by me!). The previewer supports only `f` and `w`, as it does
not include a dual-page display option.

This fixes #30. This was previously attempted in #57, but the
implementation was flawed: it neglected to disable the accelerators when
search was active, and as a consequence it was not possible to search
for strings containing `f`, `w`, or `d`.

To implement this, I changed slightly the logic of the callbacks for the
fit-page, fill-window, and dual-page actions. These actions used to use
the GUI as the source of truth; now they use the model as the source of
truth. (This is necessary because the GUI items for the toggle buttons
are not (easily) accessible when invoked from an accelerator context.
Furthermore, separating the source of truth from the UI generally proves
to be a Good Idea, anyway.)

In places where I modified existing code, I brought it up to date with
the [style guide] (in particular, replacing tabs with four spaces for
indentation).

[style guide]: https://linuxmint.gitbooks.io/developer-guidelines/content/coding_style.html#no-tabs

Test Plan:
Test that the `f`, `w`, and `d` accelerators cause the appropriate
action to be taken, and also result in the menu items' toggled state
updating to the correct value (e.g., after you press `f`, the "Best Fit"
menu item is checked, and when you press `f` again it is unchecked).
Test that you can search (<C-f> or Slash) for strings starting with
and/or containing the characters `f`, `w`, or `d` (and that the
accelerators do not trigger when you press these keys). Further test
that the same behavior holds in the previewer (except for dual-page
display and search, which the preview does not support).

Note that you should invoke the previewer as
`$INSTALL/bin/xreader-previewer` (where `$INSTALL` is your install
directory)  rather than `$INSTALL/bin/xreader -w`, for the latter will
use the `xreader-previewer` on your system path, which is likely not the
same as the development target.

In no cases are there additional warnings on the console.

wchargin-branch: add-f-w-d-accelerators