Fix Pdf backends+popup menu functionnality for sidebar annotations
Remove some line code in PDF backend
Add functionnality for popupmenu in sidebar annotations to remove
annotations or show properties for annotations.
Remove some trailing spaces, correction of alignment of the code
Fix a bug owing to the annotation cache (when an user removed every
annotation on a page and add after that he couldn't click on the new
annotations.