Fix mimetype handling and backend selection:
- Attempt to use the most specific backend for a given document.

Depending upon the order in which the backend modules were loaded
at runtime, attempting to load a jpg file (for example) could result
in a comparison of "is image/tiff an image/*" and the tiff handler
failing to open a jpg file.

- It seems as though g_content_type_is_a() was being used backwards.
  The check should be "more specific type" is a "less specific type".
  Also it expects 'content types' which could be different than mime
  types.
- Add an exact equality check before the 'is a' check to allow a
  more specific handler to handle a more specific type, before trying
  to match to a less specific handler.

So now we can get:
image/tiff equals image/tiff
image/png is a image/*
image/png is a image/*
image/png is a image/*
application/pdf equals application/pdf
image/png is a image/*
image/jpeg is a image/*

The tiff handler will handle tiff files specifically, while the image/*
handler will attempt to handle other image types.