Revert "Implements ticket #279. Added Go To Page By Number to menu"
This reverts commit 711e6820f879a1adb349966f0348729d29830705.

It's really weird to have a menu action just focus a widget that's already visible..
it looks like it's doing nothing.

If we want to make a shortcut more obvious we do so via a Help -> Shortcuts page,
not like this.

The page widget is self-explanatory, it's clearly visible and has a tooltip to
explain what it does. Not only that but the sidebar also is able to show pages,
so there's a correlation with the numbers.

If we absolutely want to add a Go to page feature in the menu, then we should do
it like this instead:

- Name it "Go to page number..." (the dots indicate more information will be asked)
- When it's activated, show a dialog to select the page number
- Click OK to go to the chosen page