Convert parse_mimetypes from returning g_char** to void
`parse_mimetypes` was supposed to return a g_char**, but it actually just stores the result in the global variable `parse_mimetypes`. To make this clearer, this commit modifies the return type of `parse_mimetypes` to be void, as the function never actually returned anything. This prevents undefined behavior in case anyone would actually try to store the returned value.
Also, the unused variable result has been removed and the indentation was fixed in one line.
This fixes https://github.com/linuxmint/xreader/issues/442