Remove landing view
A dedicated xapp called thingy is taking care of that now.
https://github.com/linuxmint/thingy

Other than the fact that thingy will do more and be faster to develop, this also fixes
a huge performance issue.

The xreader landing page actually opened each document (i.e. loaded all of its content)
just to generate a thumbnail. With a few mangas in the recent list of opened documents
this meant xreader took up to 4 or 5 minutes to actually load when no URI was specified.

Thingy just relies on Gio to load thumbs and opens instantly.