Revert "build/i18n: migrate from intltool to gettext (#552)"
This reverts commit bbac5966d1de91452033b3800c0a29a24e3a31a5.

makepot is used during releases by the Mint team only, it doesn't
nor shouldn't require any build.

This is also overly complicated and tied to the build system. If
the only issue is to translate the backend strings, then we can
fix that in the UI. I assume it's only used in the open file
dialog when changing the formats.