ev-window.c: Don't close the window before a save operation completes.
When attempting to close a window with a modified document, xreader
asks if you want to save or save a copy.

Even if you clicked save/save-as, the window (and app) would close
before the actual save could occur. As a result, the only way to
save your work would be to explicitly do so from the file menu.

The window will now be blocked from closing until a save operation
finishes, making sure nothing is lost.