ev-window.c: Don't create the epub webview until the doc type is known.
The EvWebView widget was being created and model set, only
to be destroyed later without being used if it turned out not
to be an epub document.

The model, when updated once the document was fully loaded, was
notifying the EvWebView as well as the EvView, with the former
throwing exceptions if the document wasn't compatible.